+++ NEWS.md
@@ -1,6 +1,7 @@
 # derivmkts 0.2.1
 
-* Binompopt
+* Binomopt
+
 	* Fixed: default dn=1.5 in binomopt and binomplot
 	* Added "returnprice" parameter to binomplot
 	
@@ -10,8 +11,12 @@
 
 * callperpetual and putperpetual functions added to barriers.R
 
-* Added geomavgpricecall, geomavgpriceput, geomavgstrikecall, and
-  geomavgstrikeput functions. 
+* Asian options 
+
+	* Added geomavgpricecall, geomavgpriceput, geomavgstrikecall, and
+	  geomavgstrikeput functions.
+	* Fixed greeks functionality for asian options
+	
 
 # derivmkts 0.2.0